feat(): abstract useOwnedBeepTbaDeployedStatus
